requirements.txt
asgiref==3.7.2
Django==5.0.2
django-appconf==1.0.6
django-compressor==4.4
djangorestframework==3.14.0
pytz==2024.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
